+++ src/components/Button/Button.tsx
@@ -39,10 +39,6 @@ export interface IButtonProps {
 /**
  * A button.
  * @accessibility
- * Default behavior: ButtonBehavior
- *  - adds role='button' if element type is other than 'button'
- *
- *
  * Other considerations:
  *  - for disabled buttons, add 'disabled' attribute so that the state is properly recognized by the screen reader
  *  - if button includes icon only, textual representation needs to be provided by using 'title', 'aria-label', or 'aria-labelledby' attributes
@@ -151,6 +147,7 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
     classes,
     accessibility,
     variables,
+    styles,
     rest,
   }): React.ReactNode {
     const { children, content, disabled, iconPosition } = this.props
@@ -166,19 +163,21 @@ class Button extends UIComponent<Extendable<IButtonProps>, any> {
         {...rest}
       >
         {hasChildren && children}
-        {!hasChildren && iconPosition !== 'after' && this.renderIcon()}
+        {!hasChildren && iconPosition !== 'after' && this.renderIcon(variables, styles)}
         {!hasChildren && content && <span className={classes.content}>{content}</span>}
-        {!hasChildren && iconPosition === 'after' && this.renderIcon()}
+        {!hasChildren && iconPosition === 'after' && this.renderIcon(variables, styles)}
       </ElementType>
     )
   }
 
-  public renderIcon = () => {
+  public renderIcon = (variables, styles) => {
     const { icon, iconPosition, content } = this.props
 
     return Icon.create(icon, {
       defaultProps: {
+        styles: { root: styles.icon },
         xSpacing: !content ? 'none' : iconPosition === 'after' ? 'before' : 'after',
+        variables: variables.icon,
       },
     })
   }