+++ test/specs/commonTests/isConformant.tsx
@@ -9,6 +9,7 @@ import helpers from './commonHelpers'
 
 import * as stardust from 'src/'
 import { felaRenderer } from 'src/lib'
+import { FocusZone } from 'src/lib/accessibility/FocusZone'
 
 export const mount = (node, options?) => {
   return enzymeMount(
@@ -31,12 +32,19 @@ export default (Component, options: any = {}) => {
 
   const componentType = typeof Component
 
-  // This is added because of the FelaTheme wrapper and the component itself, because it is mounted
+  // This is added because the component is mounted
   const getComponent = wrapper => {
-    return wrapper
+    // FelaTheme wrapper and the component itself:
+    let component = wrapper
       .childAt(0)
       .childAt(0)
       .childAt(0)
+    if (component.type() === FocusZone) {
+      component = component // <FocusZone> wrap
+        .childAt(0) // <div> inside FocusZone wrap
+        .childAt(0) // the actual component
+    }
+    return component
   }
 
   // make sure components are properly exported
@@ -338,7 +346,7 @@ export default (Component, options: any = {}) => {
         let errorMessage = 'was not called with (event)'
 
         if (_.has(Component.propTypes, listenerName)) {
-          expectedArgs = [eventShape, component.props()]
+          expectedArgs = [eventShape, expect.objectContaining(component.props())]
           errorMessage =
             'was not called with (event, data).\n' +
             `Ensure that 'props' object is passed to '${listenerName}'\n` +
@@ -366,6 +374,16 @@ export default (Component, options: any = {}) => {
   // Handles className
   // ----------------------------------------
   describe('static className (common)', () => {
+    const getClassesOfRootElement = component => {
+      const classes = component
+        .find('[className]')
+        .hostNodes()
+        .filterWhere(c => !c.prop('data-focuszone-id')) // filter out FocusZone <div>
+        .first()
+        .prop('className')
+      return classes
+    }
+
     test(`is a static equal to "${info.componentClassName}"`, () => {
       expect(Component.className).toEqual(info.componentClassName)
     })
@@ -376,14 +394,7 @@ export default (Component, options: any = {}) => {
       // only test components that implement className
       if (component.find('[className]').hostNodes().length > 0) {
         expect(
-          _.includes(
-            component
-              .find('[className]')
-              .hostNodes()
-              .first()
-              .prop('className'),
-            `${info.componentClassName}`,
-          ),
+          _.includes(getClassesOfRootElement(component), `${info.componentClassName}`),
         ).toEqual(true)
       }
     })
@@ -412,26 +423,13 @@ export default (Component, options: any = {}) => {
         document.body.removeChild(mountNode)
       } else {
         const component = mount(<Component {...requiredProps} className={className} />)
-        expect(
-          _.includes(
-            component
-              .find('[className]')
-              .hostNodes()
-              .first()
-              .prop('className'),
-            className,
-          ),
-        ).toEqual(true)
+        expect(_.includes(getClassesOfRootElement(component), className)).toEqual(true)
       }
     })
 
     test("user's className does not override the default classes", () => {
       const component = mount(<Component {...requiredProps} />)
-      const defaultClasses = component
-        .find('[className]')
-        .hostNodes()
-        .first()
-        .prop('className')
+      const defaultClasses = getClassesOfRootElement(component)
 
       if (!defaultClasses) return
 
@@ -439,11 +437,7 @@ export default (Component, options: any = {}) => {
       const wrapperWithCustomClasses = mount(
         <Component {...requiredProps} className={userClasses} />,
       )
-      const mixedClasses = wrapperWithCustomClasses
-        .find('[className]')
-        .hostNodes()
-        .first()
-        .prop('className')
+      const mixedClasses = getClassesOfRootElement(wrapperWithCustomClasses)
 
       const message = [
         'Make sure you are using the `getUnhandledProps` util to spread the `rest` props.',