+++ test/specs/components/Button/Button-test.tsx
@@ -13,8 +13,6 @@ import { ToggleButtonBehavior } from '../../../../src/lib/accessibility'
 import Button from 'src/components/Button/Button'
 import Icon from 'src/components/Icon/Icon'
 
-import { MenuBehavior } from 'src/lib/accessibility'
-
 const buttonImplementsShorthandProp = implementsShorthandProp(Button)
 
 jest.mock('what-input', () => {
@@ -33,8 +31,6 @@ describe('Button', () => {
     describe('button', () => {
       handlesAccessibility(Button, {
         defaultRootRole: undefined,
-        accessibilityOverride: MenuBehavior,
-        overriddenRootRole: 'menu',
       })
     })
 
@@ -42,8 +38,6 @@ describe('Button', () => {
       handlesAccessibility(Button, {
         requiredProps: { as: 'div' },
         defaultRootRole: 'button',
-        accessibilityOverride: MenuBehavior,
-        overriddenRootRole: 'menu',
       })
     })