+++ CHANGELOG.md
@@ -17,6 +17,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Fixes
+- Fix `Popup` component imports @mnajdova ([#222](https://github.com/stardust-ui/react/pull/221))
+- Fix font size issue in docs @kuzhelov ([#221](https://github.com/stardust-ui/react/pull/221))
+
 ### Features
 - Add `FocusZone` to `renderComponent`, change `Menu` behavior to support arrow keys @tomasiser ([#116](https://github.com/stardust-ui/react/pull/116))